Extract rinble-ui custom theme
