+++ frontend/packages/react-app/src/components/atoms/AddNewToken/index.stories.tsx
@@ -16,5 +16,5 @@ const Template: Story<AddNewTokenProps> = (args) => (
 
 export const Default = Template.bind({});
 Default.args = {
-  color: "itblue",
+  color: "itred",
 };
