+++ frontend/packages/react-app/src/theme/rinble-ui-theme.js
@@ -1,0 +1,20 @@
+import { theme } from "rimble-ui";
+
+const customTheme = {
+  ...theme,
+};
+
+customTheme.colors.primary = "#E25E89";
+customTheme.colors.black = "#333";
+
+customTheme.colors.ired = "#FFACC7";
+customTheme.colors.igreen = "#CAFC74";
+customTheme.colors.iblue = "#92DDE6";
+customTheme.colors.iyellow = "#EFEC6B";
+
+customTheme.colors.itred = "#E25E89";
+customTheme.colors.itgreen = "#9DD837";
+customTheme.colors.itblue = "#48C5D5";
+customTheme.colors.ityellow = "#E3DF2D";
+
+export default customTheme;
